Add Step7 and modified create GIS file
Added step7 component and remove unnecessary codes and functions in create.jsx
